# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk aplikasi keuangan Finvera - Personal Finance Manager",
        "title": "Finvera API",
        "contact": {
            "name": "Finvera Dev Team",
            "email": "dev@finvera.app"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/scheduled": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all scheduled transactions for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScheduledTransactions"
                ],
                "summary": "Get all scheduled transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new scheduled transaction for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScheduledTransactions"
                ],
                "summary": "Create a new scheduled transaction",
                "parameters": [
                    {
                        "description": "Create Scheduled Transaction Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateScheduledRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/scheduled/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get scheduled transaction details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScheduledTransactions"
                ],
                "summary": "Get scheduled transaction by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scheduled Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update scheduled transaction details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScheduledTransactions"
                ],
                "summary": "Update scheduled transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scheduled Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Scheduled Transaction Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateScheduledRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a scheduled transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScheduledTransactions"
                ],
                "summary": "Delete scheduled transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scheduled Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.CreateScheduledRequest": {
            "type": "object",
            "required": [
                "accountId",
                "amount",
                "categoryId",
                "frequency",
                "name",
                "nextRun",
                "type"
            ],
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "categoryId": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string",
                    "enum": [
                        "daily",
                        "weekly",
                        "monthly",
                        "yearly"
                    ]
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nextRun": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "targetAccountId": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense",
                        "transfer"
                    ]
                }
            }
        },
...
        "service.UpdateScheduledRequest": {
            "type": "object",
            "required": [
                "accountId",
                "amount",
                "categoryId",
                "frequency",
                "name",
                "nextRun",
                "type"
            ],
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "categoryId": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string",
                    "enum": [
                        "daily",
                        "weekly",
                        "monthly",
                        "yearly"
                    ]
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nextRun": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "targetAccountId": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense",
                        "transfer"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}